monthly budget table implementation and data architecture from UI to db + firebase and vice versa

+++ app/src/main/java/com/penny/planner/ui/screens/onboarding/SetBudgetScreen.kt
@@ -215,7 +215,7 @@ fun SetBudgetScreen(
                 showLoader = true
                 viewModel.setMonthlyLimit(
                     MonthlyBudgetInfoModel(
-                        monthlyBudget = budget,
+                        monthlyBudget = budget.toDouble(),
                         safeToSpendLimit = sliderPosition.toInt()
                     )
                 )
